refactor: move mfu utils to own module
